Separated out user class
-Moved initialization of loginmanager to the init file
-User class now within models.py
-Profile uses only the user_obj variables
