+++ flaskr/routes.py
@@ -2,63 +2,25 @@ from flaskr import app, db_connect
 from flask import render_template, request, redirect, jsonify, url_for, flash,\
     session
 from .forms import LoginForm, SignUpForm
-from flask_login import LoginManager, UserMixin, login_user, logout_user, \
+from .models import User
+from flask_login import login_user, logout_user, \
     login_required, current_user
 from .db_connect import execute_query
 from werkzeug.security import generate_password_hash, check_password_hash
 
-login_manager = LoginManager()
-login_manager.init_app(app)
-
-
-# User Model for Flash-Login
-class User(UserMixin):
-    # Constructor for User Model
-    def __init__(self, username, id, active=True):
-        self.username = username
-        self.id = id
-        self.active = active
-
-    # Returns true because users are always active
-    def is_active(self):
-        return True
-
-    # Returns false because users are always not anonymous
-    def is_anonymous(self):
-        return False
-
-    def is_authenticated(self):
-        return True
-
-    def get_id(self):
-        return self.id
-
-    @login_manager.user_loader
-    def load_user(id):
-        user_id = int(id)
-        query = """SELECT id, username FROM users WHERE id = %d;""" % user_id
-        dbuser = list(execute_query(query))
-        print(dbuser)
-        if dbuser:
-            user_obj = User(username=dbuser[0][1], id=dbuser[0][0])
-            print(type(user_obj))
-            return user_obj
-        else:
-            return None
-
-
 # Route to the login page
-@app.route('/', methods=('GET', 'POST'))
+@app.route('/login', methods=('GET', 'POST'))
 def login():
     form = LoginForm()
     if form.validate_on_submit():
         user = execute_query(
-            """SELECT id, username, password FROM users
+            """SELECT id, username, f_name, l_name, email, password FROM users
                 WHERE username = \'%s\';""" % form.username.data)
         user_list = list(user)
         if user_list:
-            if check_password_hash(user_list[0][2], form.password.data):
-                user_obj = User(username=user_list[0][1], id=user_list[0][0])
+            if check_password_hash(user_list[0][5], form.password.data):
+                user_obj = User(id=user_list[0][0], username=user_list[0][1], f_name=user_list [0][2], l_name=user_list [0][3],
+                email=user_list[0][4], password=user_list[0][5])
                 login_user(user_obj)
                 return redirect('/profile')
             else:
@@ -79,22 +41,20 @@ def signup():
         if user:
             flash("There is already an account with that name.")
         else:
+            hashed_password = generate_password_hash(form.password.data)
             query = """INSERT INTO users
                         (username, f_name, l_name, email, password)
                         VALUES (\'%s\', \'%s\', \'%s\', \'%s\', \'%s\')""" \
                     % (form.username.data, form.f_name.data,
-                       form.l_name.data, form.email.data,
-                       generate_password_hash(
-                           form.password.data))
+                       form.l_name.data, form.email.data, hashed_password)
             try:
                 insert = execute_query(query)
                 return_id = list(
                     execute_query("""SELECT id FROM users
                                       WHERE username = \'%s\';"""
                                   % form.username.data))
-                user_obj = User(username=form.username.data,
-                                id=return_id[0][0])
-
+                user_obj = User(id=return_id[0][0], username=form.username.data, f_name=form.f_name.data, l_name=form.l_name.data,
+                email=form.email.data, password=hashed_password)
                 login_user(user_obj)
                 flash("You have successfully signed up!")
                 return redirect('/profile')
@@ -108,17 +68,7 @@ def signup():
 @app.route('/profile')
 @login_required
 def profile():
-    user_id = current_user.get_id()
-    try:
-        user = list(execute_query(
-            """SELECT username, f_name, l_name, email FROM users
-            WHERE id = %d;""" % user_id))
-
-        return render_template('profile.html', profile=user[0])
-
-    except:
-        flash("Error")
-        return render_template('404.html')
+    return render_template('profile.html')
 
 
 @app.route('/logout')
@@ -126,7 +76,7 @@ def profile():
 def logout():
     logout_user()
     flash("Account has been logged out")
-    return redirect('/login')
+    return redirect('/')
 
 
 # Route for guest (restrict saving recipes)
@@ -388,7 +338,7 @@ def add_ingredients():
         return redirect(url_for('recipe_display'))
 
 
-@app.route('/homepage')
+@app.route('/')
 def homepage():
     return render_template('homepage.html')
