+++ flaskr/templates/base.html
@@ -11,21 +11,22 @@
 	<body>
 		<h1>{% block h1 %}{% endblock %}</h1>
 		<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-  		<a class="navbar-brand" href="#">Ethical Eating</a>
+  		<a class="navbar-brand" href="/">Ethical Eating</a>
   		<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
     		<span class="navbar-toggler-icon"></span>
   		</button>
   		<div class="collapse navbar-collapse" id="navbarNavDropdown">
     		<ul class="navbar-nav justify-content-center">
       		<li class="nav-item active">
-        		<a class="nav-link" href="/homepage">Home<span class="sr-only">(current)</span></a>
+        		<a class="nav-link" href="/">Home<span class="sr-only">(current)</span></a>
       		</li>
 					<li class="nav-item dropdown">
         		<a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
     				Login
         		</a>
         		<div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-          		<a class="dropdown-item" href="/">Sign In</a>
+          		<a class="dropdown-item" href="/login">Sign In</a>
+							<a class="dropdown-item" href="/profile">My Profile</a>
           		<a class="dropdown-item" href="/signup">Sign Up</a>
           		<a class="dropdown-item" href="/logout">Logout</a>
         		</div>
@@ -42,7 +43,7 @@
       		<li class="nav-item">
         		<a class="nav-link" href="/add_ingredients">Ingredients</a>
       		</li>
-      		<li class="nav-item">  
+      		<li class="nav-item">
         		<a class="nav-link" href="/faq">FAQ</a>
       		</li>
     		</ul>
