+++ flaskr/templates/profile.html
@@ -1,7 +1,18 @@
 {% extends 'base.html' %}
+
+
+
+
+
 {% block content  %}
-<p>{{ profile.0}}</p>
-<p>{{ profile.1}}</p>
-<p>{{ profile.2}}</p>
-<p>{{ profile.3}}</p>
+
+{% if current_user.is_authenticated %}
+  <p>Id: {{ current_user.id }}<p>
+  <p>Username: {{ current_user.username }}<p>
+  <p>First Name: {{ current_user.f_name }}<p>
+  <p>Last Name: {{ current_user.l_name }}<p>
+  <p>Email: {{ current_user.email }}<p>
+  <p>Password: {{ current_user.password }}<p>
+{% endif %}
+
 {% endblock %}
